refactor(hush-cli): extract audit_stats from pkg_cli
Move cmd_pkg_audit and cmd_pkg_stats into pkg_cli/audit_stats.rs, both
widened to pub(super). format_number/urlencoding_simple pulled from
super::util; format_number dropped from mod.rs's util import (its only
consumer moved here).
